edit recipe done

+++ src/components/Home.jsx
@@ -1,13 +1,16 @@
 import { useState, useEffect } from "react";
 import AddRecipeModal from "./AddRecipeModal";
 import RecipeDetail from "./RecipeDetail";
-import { AiOutlinePlus, AiOutlineDelete } from "react-icons/ai";
+import { AiOutlinePlus } from "react-icons/ai";
+import { toast } from "react-toastify";
 
 const Home = () => {
   const [isOpen, setIsOpen] = useState(false);
   const [recipeData, setRecipeData] = useState([]);
 
   const [selectedId, setSelectedId] = useState(null);
+
+  const [editId, setEditId] = useState(null);
 
   const handleAdd = () => {
     setIsOpen(true);
@@ -17,16 +20,38 @@ const Home = () => {
   };
 
   const handleSave = (recipe) => {
-    const dataToSet = [...recipeData, recipe];
-    setRecipeData(dataToSet);
-    localStorage.setItem("recipe", JSON.stringify(dataToSet));
+    const existingRecipe = recipeData.find((r) => +r.id === +recipe.id);
+  
+    if (existingRecipe) {
+      const updateData = recipeData.map((r) =>
+        r.id === recipe.id ? { ...r, ...recipe } : r
+      );
+      setRecipeData(updateData);
+      localStorage.setItem("recipe", JSON.stringify(updateData));
+    } else {
+      const newData = [...recipeData, recipe];
+      setRecipeData(newData);
+      localStorage.setItem("recipe", JSON.stringify(newData));
+    }
   };
 
   const handleDelete = (id) => {
-    const updatedRecipeData = recipeData.filter((data) => data.id !== id);
-    setRecipeData(updatedRecipeData);
-    // setSelectedId(id - 1);
-    localStorage.setItem("recipe", JSON.stringify(updatedRecipeData));
+    const confirmDelete = confirm("Are you sure you want to delete the recipe");
+    if (confirmDelete) {
+      const updatedRecipeData = recipeData.filter((data) => data.id !== id);
+      setRecipeData(updatedRecipeData);
+      localStorage.setItem("recipe", JSON.stringify(updatedRecipeData));
+      toast.success("Recipe deleted successfully");
+    }
+  };
+
+  const handleEdit = (id) => {
+    if (!id) {
+      toast.error("id not provided");
+      return;
+    }
+    setIsOpen(true);
+    setEditId(id);
   };
 
   useEffect(() => {
@@ -64,22 +89,11 @@ const Home = () => {
                         } p-2 rounded hover:cursor-pointer hover:bg-gray-300 `}
                         onClick={() => handleTitleClick(recipe.id)}
                       >
-                        <span className=" flex justify-between">
-                          <span className=" uppercase">
-                            {"->"}{" "}
-                            {recipe.title && recipe.title.length > 30
-                              ? recipe.title.slice(0, 30) + "..."
-                              : recipe.title || "Untitled Recipe"}
-                          </span>
-                          <span
-                            onClick={(e) => {
-                              e.stopPropagation();
-                              handleDelete(recipe.id);
-                            }}
-                            className=" hover:text-red-600"
-                          >
-                            <AiOutlineDelete className=" text-xl" />
-                          </span>
+                        <span className=" uppercase">
+                          {"->"}{" "}
+                          {recipe.title && recipe.title.length > 30
+                            ? recipe.title.slice(0, 30) + "..."
+                            : recipe.title || "Untitled Recipe"}
                         </span>
                       </div>
                     </div>
@@ -91,16 +105,29 @@ const Home = () => {
             {selectedId && (
               <RecipeDetail
                 data={recipeData.filter((obj) => obj.id === selectedId)}
+                handleDelete={handleDelete}
+                handleEdit={handleEdit}
               />
             )}
           </div>
-          <AddRecipeModal
-            isOpen={isOpen}
-            onClose={handleClose}
-            onSave={handleSave}
-            recipeData={recipeData}
-            setRecipeData={setRecipeData}
-          />
+          {editId ? (
+            <AddRecipeModal
+              isOpen={isOpen}
+              onClose={handleClose}
+              onSave={handleSave}
+              recipeData={recipeData}
+              updateData={recipeData.filter((obj) => obj.id === editId)}
+              setRecipeData={setRecipeData}
+            />
+          ) : (
+            <AddRecipeModal
+              isOpen={isOpen}
+              onClose={handleClose}
+              onSave={handleSave}
+              recipeData={recipeData}
+              setRecipeData={setRecipeData}
+            />
+          )}
         </div>
       ) : (
         <div className=" ">
